Move Batcher out into its own impl module.
